requirements.txt
loguru==0.6.0
numpy==1.23.5
pandas==1.4.2
PyYAML==6.0
scikit_image==0.19.2
scikit_multilearn==0.2.0
ultralytics==8.0.9
